Move GetDefaultBranch to gitserver client
fix mocks
