+++ cmd/frontend/graphqlbackend/repository_test.go
@@ -18,7 +18,6 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain"
 	"github.com/sourcegraph/sourcegraph/internal/search/result"
 	"github.com/sourcegraph/sourcegraph/internal/types"
-	"github.com/sourcegraph/sourcegraph/internal/vcs/git"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
 
@@ -206,11 +205,11 @@ func TestRepository_DefaultBranch(t *testing.T) {
 	}
 	for _, tt := range ts {
 		t.Run(tt.name, func(t *testing.T) {
-			git.Mocks.ExecSafe = func(params []string) (stdout []byte, stderr []byte, exitCode int, err error) {
+			gitserver.Mocks.ExecSafe = func(params []string) (stdout []byte, stderr []byte, exitCode int, err error) {
 				return []byte(tt.symbolicRef), nil, tt.symbolicRefExitCode, tt.symbolicRefErr
 			}
 			t.Cleanup(func() {
-				git.Mocks.ExecSafe = nil
+				gitserver.Mocks.ExecSafe = nil
 			})
 
 			gitserver.Mocks.ResolveRevision = func(spec string, opt gitserver.ResolveRevisionOptions) (api.CommitID, error) {
