+++ cmd/frontend/graphqlbackend/search_results_stats_languages_test.go
@@ -40,7 +40,7 @@ func TestSearchResultsStatsLanguages(t *testing.T) {
 		return io.NopCloser(bytes.NewReader(data)), nil
 	}
 	const wantDefaultBranchRef = "refs/heads/foo"
-	git.Mocks.ExecSafe = func(params []string) (stdout, stderr []byte, exitCode int, err error) {
+	gitserver.Mocks.ExecSafe = func(params []string) (stdout, stderr []byte, exitCode int, err error) {
 		// Mock default branch lookup in (*RepositoryResolver).DefaultBranch.
 		return []byte(wantDefaultBranchRef), nil, 0, nil
 	}
